ENH: rework global marker handling
Include markers 3, 4 in the global marker pool if they connect.
Indicate clearly which markers are global and which are local,
this is helpful if the IOC doesn't have all 4.
Improve performance of marker code, doesn't slow down the GUI as much.
Remove vestigial code for an older version that had weird special
global 2 mode for markers 3, 4 with different behavior.